Paramertize data-model & move to suites.
